Refactor API to use an interface.
Interfaces allow mocking libraries to mock our types more easily. This
will prevent users of KVault from having to wrap KVault with another
type when they want to test code which uses it.
